+++ modules/utils/__init__.py
@@ -9,7 +9,6 @@ from .notification import *
 from .data_utils import *
 from .I18nstr import *
 from .baah_exceptions import *
-from .download_utils import *
 
 from modules.utils.log_utils import logging
 import time
